move bech32AssetId to getters to address circular dependency
