+++ packages/router/package.json
@@ -16,5 +16,8 @@
     "@penumbra-zone/storage": "workspace:*",
     "@penumbra-zone/transport-dom": "workspace:*",
     "@penumbra-zone/wasm": "workspace:*"
+  },
+  "devDependencies": {
+    "@penumbra-zone/getters": "workspace:*"
   }
 }
