+++ packages/router/src/grpc/custody/authorize.ts
@@ -2,7 +2,8 @@ import type { Impl } from '.';
 import { approverCtx, extLocalCtx, extSessionCtx, servicesCtx } from '../../ctx';
 import { authorizePlan, generateSpendKey } from '@penumbra-zone/wasm';
 import { Key } from '@penumbra-zone/crypto-web';
-import { bech32AssetId, Box, Jsonified } from '@penumbra-zone/types';
+import { bech32AssetId } from '@penumbra-zone/getters/asset';
+import { Box, Jsonified } from '@penumbra-zone/types';
 import { Code, ConnectError, HandlerContext } from '@connectrpc/connect';
 import { Metadata } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
 import { viewTransactionPlan } from './view-transaction-plan';
