+++ packages/types/package.json
@@ -14,5 +14,8 @@
     "bignumber.js": "^9.1.2",
     "idb": "^8.0.0",
     "zod": "^3.22.4"
+  },
+  "devDependencies": {
+    "@penumbra-zone/getters": "workspace:*"
   }
 }
