+++ packages/types/src/index.ts
@@ -8,7 +8,6 @@ export * from './environment';
 export * from './box';
 export * from './wallet';
 export * from './utility';
-export * from './asset';
 export * from './transaction';
 export * from './translators';
 export * from './hex';
@@ -22,7 +21,6 @@ export * from './chain';
 export * from './jsonified';
 export * from './schemas';
 export * from './type-predicates';
-export * from './getters';
 export * from './staking';
 export * from './identity-key';
 export * from './customize-symbol';
