+++ packages/ui/package.json
@@ -40,6 +40,7 @@
     "tinycolor2": "^1.6.0"
   },
   "devDependencies": {
+    "@penumbra-zone/getters": "workspace:*",
     "@storybook/addon-essentials": "^7.6.17",
     "@storybook/addon-interactions": "^7.6.17",
     "@storybook/addon-links": "^7.6.17",
